@@ -28,9 +28,11 @@
   },
   "homepage": "https://github.com/pixely/365-photos#readme",
   "dependencies": {
-    "babel-cli": "^6.24.0",
     "babel-polyfill": "6.8.0",
     "babel-preset-node6": "^11.0.0",
+    "color-thief": "lokesh/color-thief",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
     "normalize.css": "^5.0.0",
     "react": "15.0.2",
     "react-dom": "15.0.2",
@@ -42,7 +44,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^6.3.6",
-    "babel-cli": "^6.8.0",
+    "babel-cli": "^6.24.1",
     "babel-core": "6.8.0",
     "babel-loader": "6.2.4",
     "babel-plugin-react-display-name": "2.0.0",
@@ -56,10 +58,10 @@
     "cross-env": "1.0.7",
     "css-loader": "0.23.1",
     "enzyme": "2.2.0",
-    "eslint": "2.9.0",
-    "eslint-plugin-import": "1.6.1",
-    "eslint-plugin-react": "5.0.1",
-    "eslint-watch": "2.1.11",
+    "eslint": "^3.19.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-react": "^6.10.3",
+    "eslint-watch": "^3.0.1",
     "eventsource-polyfill": "0.9.6",
     "expect": "1.19.0",
     "express": "4.13.4",
