@@ -2,56 +2,65 @@ import React from 'react';
 import TextInput from '../common/TextInput';
 import SelectInput from '../common/SelectInput';
 
-const CourseForm = ({course, allAuthors, onSave, onChange, saving, errors}) => {
-  return (
-    <form>
-      <h1>Manage Course</h1>
-      <TextInput
-        name="title"
-        label="Title"
-        value={course.title}
-        onChange={onChange}
-        error={errors.title}/>
+const CourseForm = ({ course, allAuthors, onSave, onChange, saving, errors }) => (
+  <form>
+    <h1>Manage Course</h1>
+    <TextInput
+      name="title"
+      label="Title"
+      value={course.title}
+      onChange={onChange}
+      error={errors.title}
+    />
 
-      <SelectInput
-        name="authorId"
-        label="Author"
-        value={course.authorId}
-        defaultOption="Select Author"
-        options={allAuthors}
-        onChange={onChange} error={errors.authorId}/>
+    <SelectInput
+      name="authorId"
+      label="Author"
+      value={course.authorId}
+      defaultOption="Select Author"
+      options={allAuthors}
+      onChange={onChange} error={errors.authorId}
+    />
 
-      <TextInput
-        name="category"
-        label="Category"
-        value={course.category}
-        onChange={onChange}
-        error={errors.category}/>
+    <TextInput
+      name="category"
+      label="Category"
+      value={course.category}
+      onChange={onChange}
+      error={errors.category}
+    />
 
-      <TextInput
-        name="length"
-        label="Length"
-        value={course.length}
-        onChange={onChange}
-        error={errors.length}/>
+    <TextInput
+      name="length"
+      label="Length"
+      value={course.length}
+      onChange={onChange}
+      error={errors.length}
+    />
 
-      <input
-        type="submit"
-        disabled={saving}
-        value={saving ? 'Saving...' : 'Save'}
-        className="btn btn-primary"
-        onClick={onSave}/>
-    </form>
-  );
+    <input
+      type="submit"
+      disabled={saving}
+      value={saving ? 'Saving...' : 'Save'}
+      className="btn btn-primary"
+      onClick={onSave}
+    />
+  </form>
+);
+
+CourseForm.defaultProps = {
+  allAuthors: [],
+  saving: false,
+  errors: {},
 };
 
 CourseForm.propTypes = {
-  course: React.PropTypes.object.isRequired,
-  allAuthors: React.PropTypes.array,
+  course: React.PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
+  allAuthors: React.PropTypes.array, // eslint-disable-line react/forbid-prop-types
   onSave: React.PropTypes.func.isRequired,
   onChange: React.PropTypes.func.isRequired,
   saving: React.PropTypes.bool,
-  errors: React.PropTypes.object
+  errors: React.PropTypes.object, // eslint-disable-line react/forbid-prop-types
 };
 
 export default CourseForm;
