@@ -1,18 +1,18 @@
 import * as types from './actionTypes';
 import authorApi from '../api/mockAuthorApi';
-import {beginAjaxCall} from './ajaxStatusActions';
+import { beginAjaxCall } from './ajaxStatusActions';
 
 export function loadAuthorsSuccess(authors) {
   return { type: types.LOAD_AUTHORS_SUCCESS, authors };
 }
 
 export function loadAuthors() {
-  return function(dispatch) {
+  return (dispatch) => {
     dispatch(beginAjaxCall());
-    return authorApi.getAllAuthors().then(authors => {
+    return authorApi.getAllAuthors().then((authors) => {
       dispatch(loadAuthorsSuccess(authors));
-    }).catch(error => {
-      throw(error);
+    }).catch((error) => {
+      throw (error);
     });
   };
 }
