@@ -1,6 +1,6 @@
 import * as types from './actionTypes';
 import courseApi from '../api/mockCourseApi';
-import {beginAjaxCall, ajaxCallError} from './ajaxStatusActions';
+import { beginAjaxCall, ajaxCallError } from './ajaxStatusActions';
 
 export function loadCoursesSuccess(courses) {
   return { type: types.LOAD_COURSES_SUCCESS, courses };
@@ -15,25 +15,26 @@ export function updateCourseSuccess(course) {
 }
 
 export function loadCourses() {
-  return function(dispatch) {
+  return (dispatch) => {
     dispatch(beginAjaxCall());
-    return courseApi.getAllCourses().then(courses => {
+    return courseApi.getAllCourses().then((courses) => {
       dispatch(loadCoursesSuccess(courses));
-    }).catch(error => {
-      throw(error);
+    }).catch((error) => {
+      throw (error);
     });
   };
 }
 
 export function saveCourse(course) {
-  return function(dispatch, getState) {
+  return (dispatch, getState) => {
     dispatch(beginAjaxCall());
-    return courseApi.saveCourse(course).then(savedCourse => {
+    return courseApi.saveCourse(course).then((savedCourse) => {
+      // eslint-disable-next-line no-unused-expressions
       course.id ? dispatch(updateCourseSuccess(savedCourse)) :
         dispatch(createCourseSuccess(savedCourse));
-    }).catch(error => {
+    }).catch((error) => {
       dispatch(ajaxCallError());
-      throw(error);
+      throw (error);
     });
   };
 }
