@@ -1,26 +1,25 @@
 import expect from 'expect';
-import * as courseActions from './courseActions';
-import * as types from './actionTypes';
-
 import thunk from 'redux-thunk';
 import nock from 'nock';
 import configureMockStore from 'redux-mock-store';
+import * as courseActions from './courseActions';
+import * as types from './actionTypes';
 
 // Test a sync action
 describe('Course Actions', () => {
   describe('createCourseSuccess', () => {
     it('should create a CREATE_COURSE_SUCCESS action', () => {
-      //arrange
-      const course = {id: 'clean-code', title: 'Clean Code'};
+      // arrange
+      const course = { id: 'clean-code', title: 'Clean Code' };
       const expectedAction = {
         type: types.CREATE_COURSE_SUCCESS,
-        course: course
+        course,
       };
 
-      //act
+      // act
       const action = courseActions.createCourseSuccess(course);
 
-      //assert
+      // assert
       expect(action).toEqual(expectedAction);
     });
   });
@@ -41,11 +40,11 @@ describe('Async Actions', () => {
     //     .reply(200, {body: {course: [{id: 1, firstName: 'Cory', lastName: 'House'}] }});
 
     const expectedActions = [
-      {type: types.BEGIN_AJAX_CALL},
-      {type: types.LOAD_COURSES_SUCCESS, body: {courses:[{id: 'clean-code', title: 'Clean Code'}]}}
+      { type: types.BEGIN_AJAX_CALL },
+      { type: types.LOAD_COURSES_SUCCESS, body: { courses: [{ id: 'clean-code', title: 'Clean Code' }] } },
     ];
 
-    const store = mockStore({courses: []}, expectedActions);
+    const store = mockStore({ courses: [] }, expectedActions);
     store.dispatch(courseActions.loadCourses()).then(() => {
       const actions = store.getActions();
       expect(actions[0].type).toEqual(types.BEGIN_AJAX_CALL);
