@@ -7,24 +7,22 @@ const authors = [
   {
     id: 'cory-house',
     firstName: 'Cory',
-    lastName: 'House'
+    lastName: 'House',
   },
   {
     id: 'scott-allen',
     firstName: 'Scott',
-    lastName: 'Allen'
+    lastName: 'Allen',
   },
   {
     id: 'dan-wahlin',
     firstName: 'Dan',
-    lastName: 'Wahlin'
-  }
+    lastName: 'Wahlin',
+  },
 ];
 
-//This would be performed on the server in a real app. Just stubbing in.
-const generateId = (author) => {
-  return author.firstName.toLowerCase() + '-' + author.lastName.toLowerCase();
-};
+// This would be performed on the server in a real app. Just stubbing in.
+const generateId = author => `${author.firstName.toLowerCase()}-${author.lastName.toLowerCase()}`;
 
 class AuthorApi {
   static getAllAuthors() {
@@ -35,8 +33,8 @@ class AuthorApi {
     });
   }
 
-  static saveAuthor(author) {
-    author = Object.assign({}, author); // to avoid manipulating object passed in.
+  static saveAuthor(passedAuthor) {
+    const author = Object.assign({}, passedAuthor); // to avoid manipulating object passed in.
     return new Promise((resolve, reject) => {
       setTimeout(() => {
         // Simulate server-side validation
@@ -50,12 +48,12 @@ class AuthorApi {
         }
 
         if (author.id) {
-          const existingAuthorIndex = authors.findIndex(a => a.id == author.id);
+          const existingAuthorIndex = authors.findIndex(a => a.id === author.id);
           authors.splice(existingAuthorIndex, 1, author);
         } else {
-          //Just simulating creation here.
-          //The server would generate ids for new authors in a real app.
-          //Cloning so copy returned is passed by value rather than by reference.
+          // Just simulating creation here.
+          // The server would generate ids for new authors in a real app.
+          // Cloning so copy returned is passed by value rather than by reference.
           author.id = generateId(author);
           authors.push(author);
         }
@@ -68,9 +66,7 @@ class AuthorApi {
   static deleteAuthor(authorId) {
     return new Promise((resolve, reject) => {
       setTimeout(() => {
-        const indexOfAuthorToDelete = authors.findIndex(author => {
-          author.id == authorId;
-        });
+        const indexOfAuthorToDelete = authors.findIndex(author => author.id === authorId);
         authors.splice(indexOfAuthorToDelete, 1);
         resolve();
       }, delay);
