@@ -1,28 +1,24 @@
-import React, {PropTypes} from 'react';
+import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
 import Header from './common/Header';
-import {connect} from 'react-redux';
 
-class App extends React.Component {
-  render() {
-    return (
-      <div>
-        <Header
-          loading={this.props.loading}
-        />
-        {this.props.children}
-      </div>
-    );
-  }
-}
+const App = props => (
+  <div>
+    <Header
+      loading={props.loading}
+    />
+    {props.children}
+  </div>
+);
 
 App.propTypes = {
-  children: PropTypes.object.isRequired,
-  loading: PropTypes.bool.isRequired
+  children: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
+  loading: PropTypes.bool.isRequired,
 };
 
 function mapStateToProps(state, ownProps) {
   return {
-    loading: state.ajaxCallsInProgress > 0
+    loading: state.ajaxCallsInProgress > 0,
   };
 }
 
