@@ -1,14 +1,10 @@
 import React from 'react';
 
-class AboutPage extends React.Component {
-  render() {
-    return (
-        <div>
-          <h1>About</h1>
-          <p>This application uses React, Redux, Router and a variety of other helpful libraries.</p>
-        </div>
-    );
-  }
-}
+const AboutPage = () => (
+  <div>
+    <h1>About</h1>
+    <p>This application uses React, Redux, Router and a variety of other helpful libraries.</p>
+  </div>
+);
 
 export default AboutPage;
