@@ -1,16 +1,16 @@
-import React, {PropTypes} from 'react';
+import React, { PropTypes } from 'react';
 
 class LoadingDots extends React.Component {
   constructor(props, context) {
     super(props, context);
 
-    this.state = {frame: 1};
+    this.state = { frame: 1 };
   }
 
   componentDidMount() {
     this.interval = setInterval(() => {
       this.setState({  // eslint-disable-line react/no-did-mount-set-state
-        frame: this.state.frame + 1
+        frame: this.state.frame + 1,
       });
     }, this.props.interval);
   }
@@ -24,19 +24,20 @@ class LoadingDots extends React.Component {
     let text = '';
     while (dots > 0) {
       text += '.';
-      dots--;
+      dots -= 1;
     }
     return <span {...this.props}>{text}&nbsp;</span>;
   }
 }
 
 LoadingDots.defaultProps = {
-  interval: 300, dots: 3
+  interval: 300,
+  dots: 3,
 };
 
 LoadingDots.propTypes = {
   interval: PropTypes.number,
-  dots: PropTypes.number
+  dots: PropTypes.number,
 };
 
 export default LoadingDots;
