@@ -1,12 +1,10 @@
 import React from 'react';
-import {IndexLink} from 'react-router';
+import { IndexLink } from 'react-router';
 
-const Logo = () => {
-  return (
-    <IndexLink to="/" className="logo">
-      365<span className="logo__sub">Photos</span>
-    </IndexLink>
-  );
-};
+const Logo = () => (
+  <IndexLink to="/" className="logo">
+    365<span className="logo__sub">Photos</span>
+  </IndexLink>
+);
 
 export default Logo;
