@@ -1,26 +1,31 @@
-import React, {PropTypes} from 'react';
+import React, { PropTypes } from 'react';
 
-const SelectInput = ({name, label, onChange, defaultOption, value, error, options}) => {
-  return (
-    <div className="form-group">
-      <label htmlFor={name}>{label}</label>
-      <div className="field">
-        {/* Note, value is set here rather than on the option - docs: https://facebook.github.io/react/docs/forms.html */}
-        <select
-          name={name}
-          value={value}
-          onChange={onChange}
-          className="form-control">
-          <option value="">{defaultOption}</option>
-          {options.map((option) => {
-            return <option key={option.value} value={option.value}>{option.text}</option>;
-          })
-          }
-        </select>
-        {error && <div className="alert alert-danger">{error}</div>}
-      </div>
+const SelectInput = ({ name, label, onChange, defaultOption, value, error, options }) => (
+  <div className="form-group">
+    <label htmlFor={name}>{label}</label>
+    <div className="field">
+      {/* Note, value is set here rather than on the option - docs: https://facebook.github.io/react/docs/forms.html */}
+      <select
+        name={name}
+        value={value}
+        onChange={onChange}
+        className="form-control"
+      >
+        <option value="">{defaultOption}</option>
+        {options.map(option =>
+          <option key={option.value} value={option.value}>{option.text}</option>,
+        )}
+      </select>
+      {error && <div className="alert alert-danger">{error}</div>}
     </div>
-  );
+  </div>
+);
+
+SelectInput.defaultProps = {
+  defaultOption: null,
+  value: null,
+  error: null,
+  options: [],
 };
 
 SelectInput.propTypes = {
@@ -30,7 +35,7 @@ SelectInput.propTypes = {
   defaultOption: PropTypes.string,
   value: PropTypes.string,
   error: PropTypes.string,
-  options: PropTypes.arrayOf(PropTypes.object)
+  options: PropTypes.arrayOf(PropTypes.object), // eslint-disable-line react/forbid-prop-types
 };
 
 export default SelectInput;
