@@ -1,16 +1,18 @@
 import expect from 'expect';
 import React from 'react';
-import {mount, shallow} from 'enzyme';
+import { mount, shallow } from 'enzyme';
 import CourseForm from './CourseForm';
 
 function setup(saving) {
-  let props = {
-    course: {}, saving: saving, errors: {},
+  const props = {
+    course: {},
+    saving,
+    errors: {},
     onSave: () => {},
-    onChange: () => {}
+    onChange: () => {},
   };
 
-  return shallow(<CourseForm {...props}/>);
+  return shallow(<CourseForm {...props} />);
 }
 
 describe('CourseForm via Enzyme', () => {
