@@ -4,28 +4,30 @@ import TestUtils from 'react-addons-test-utils';
 import CourseForm from './CourseForm';
 
 function setup(saving) {
-  let props = {
-      course: {}, saving: saving, errors: {},
-      onSave: () => {},
-      onChange: () => {}
+  const props = {
+    course: {},
+    saving,
+    errors: {},
+    onSave: () => {},
+    onChange: () => {},
   };
 
-  let renderer = TestUtils.createRenderer();
-  renderer.render(<CourseForm {...props}/>);
-  let output = renderer.getRenderOutput();
+  const renderer = TestUtils.createRenderer();
+  renderer.render(<CourseForm {...props} />);
+  const output = renderer.getRenderOutput();
 
   return {
     props,
     output,
-    renderer
+    renderer,
   };
 }
 
 describe('CourseForm via React Test Utils', () => {
   it('renders form and h1', () => {
-    const {output} = setup();
+    const { output } = setup();
     expect(output.type).toBe('form');
-    let [ h1 ] = output.props.children;
+    const [h1] = output.props.children;
     expect(h1.type).toBe('h1');
   });
 
