@@ -1,29 +1,27 @@
-import React, {PropTypes} from 'react';
+import React, { PropTypes } from 'react';
 import CourseListRow from './CourseListRow';
 
-const CourseList = ({courses}) => {
-  return (
-    <table className="table">
-      <thead>
-        <tr>
-          <th>&nbsp;</th>
-          <th>Title</th>
-          <th>Author</th>
-          <th>Category</th>
-          <th>Length</th>
-        </tr>
-      </thead>
-      <tbody>
+const CourseList = ({ courses }) => (
+  <table className="table">
+    <thead>
+      <tr>
+        <th>&nbsp;</th>
+        <th>Title</th>
+        <th>Author</th>
+        <th>Category</th>
+        <th>Length</th>
+      </tr>
+    </thead>
+    <tbody>
       {courses.map(course =>
-        <CourseListRow key={course.id} course={course}/>
-      )}
-      </tbody>
-    </table>
-  );
-};
+        <CourseListRow key={course.id} course={course} />,
+       )}
+    </tbody>
+  </table>
+);
 
 CourseList.propTypes = {
-  courses: PropTypes.array.isRequired
+  courses: PropTypes.array.isRequired, // eslint-disable-line react/forbid-prop-types
 };
 
 export default CourseList;
