@@ -1,9 +1,8 @@
-import React, {PropTypes} from 'react';
-import {connect} from 'react-redux';
-import {bindActionCreators} from 'redux';
+import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 import * as courseActions from '../../actions/courseActions';
 import CourseList from './CourseList';
-import {browserHistory} from 'react-router';
 
 class CoursesPage extends React.Component {
   constructor(props, context) {
@@ -12,42 +11,43 @@ class CoursesPage extends React.Component {
   }
 
   courseRow(course, index) {
-    return <div key={index}>{course.title}</div>;
+    return <div key={index}>{ this.course.title }</div>;
   }
 
   redirectToAddCoursePage() {
-    browserHistory.push('/course');
+    this.browserHistory.push('/course');
   }
 
   render() {
-    const {courses} = this.props;
+    const { courses } = this.props;
     return (
-        <div>
-          <h1>Courses</h1>
-          <input type="submit"
-                 value="Add Course"
-                 className="btn btn-primary"
-                 onClick={this.redirectToAddCoursePage}/>
-          <CourseList courses={courses}/>
-        </div>
+      <div>
+        <h1>Courses</h1>
+        <input
+          type="submit"
+          value="Add Course"
+          className="btn btn-primary"
+          onClick={this.redirectToAddCoursePage}
+        />
+        <CourseList courses={courses} />
+      </div>
     );
   }
 }
 
 CoursesPage.propTypes = {
-  courses: PropTypes.array.isRequired,
-  actions: PropTypes.object.isRequired
+  courses: PropTypes.array.isRequired, // eslint-disable-line react/forbid-prop-types
 };
 
 function mapStateToProps(state, ownProps) {
   return {
-    courses: state.courses
+    courses: state.courses,
   };
 }
 
 function mapDispatchToProps(dispatch) {
   return {
-    actions: bindActionCreators(courseActions, dispatch)
+    actions: bindActionCreators(courseActions, dispatch),
   };
 }
 
