@@ -1,9 +1,9 @@
-import React, {PropTypes} from 'react';
-import {connect} from 'react-redux';
-import {bindActionCreators} from 'redux';
+import React, { PropTypes } from 'react';
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 import * as courseActions from '../../actions/courseActions';
 import CourseForm from './CourseForm';
-import {authorsFormattedForDropdown} from '../../selectors/selectors';
+import { authorsFormattedForDropdown } from '../../selectors/selectors';
 
 export class ManageCoursePage extends React.Component {
   constructor(props, context) {
@@ -12,7 +12,7 @@ export class ManageCoursePage extends React.Component {
     this.state = {
       course: Object.assign({}, this.props.course),
       errors: {},
-      saving: false
+      saving: false,
     };
 
     this.saveCourse = this.saveCourse.bind(this);
@@ -20,28 +20,28 @@ export class ManageCoursePage extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (this.props.course.id != nextProps.course.id) {
-      this.setState({course: Object.assign({}, nextProps.course)});
+    if (this.props.course.id !== nextProps.course.id) {
+      this.setState({ course: Object.assign({}, nextProps.course) });
     }
   }
 
   updateCourseState(event) {
     const field = event.target.name;
-    let course = this.state.course;
+    const course = this.state.course;
     course[field] = event.target.value;
-    return this.setState({course:course});
+    return this.setState({ course });
   }
 
   courseFormIsValid() {
     let formIsValid = true;
-    let errors = {};
+    const errors = {};
 
     if (this.state.course.title.length < 5) {
       errors.title = 'Title must be at least 5 characters.';
       formIsValid = false;
     }
 
-    this.setState({errors:errors});
+    this.setState({ errors });
     return formIsValid;
   }
 
@@ -52,16 +52,16 @@ export class ManageCoursePage extends React.Component {
       return;
     }
 
-    this.setState({saving:true});
+    this.setState({ saving: true });
     this.props.actions.saveCourse(this.state.course)
       .then(() => this.redirect())
-      .catch(error => {
-        this.setState({saving:false});
+      .catch((error) => {
+        this.setState({ saving: false });
       });
   }
 
   redirect() {
-    this.setState({saving:false});
+    this.setState({ saving: false });
     this.context.router.push('/courses');
   }
 
@@ -80,38 +80,38 @@ export class ManageCoursePage extends React.Component {
 }
 
 ManageCoursePage.propTypes = {
-  course: PropTypes.object.isRequired,
-  authors: PropTypes.array.isRequired,
-  actions: PropTypes.object.isRequired
+  course: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
+  authors: PropTypes.array.isRequired, // eslint-disable-line react/forbid-prop-types
+  actions: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
 };
 
 ManageCoursePage.contextTypes = {
-  router: PropTypes.object
+  router: PropTypes.object,
 };
 
 function getCourseById(courses, id) {
-  const course = courses.filter(course => course.id == id);
-  if (course.length) return course[0];
+  const newCourse = courses.filter(course => course.id === id);
+  if (newCourse.length) return newCourse[0];
   return null;
 }
 
 function mapStateToProps(state, ownProps) {
   const courseId = ownProps.params.id;
-  let course = {id: '', watchHref: '', title: '', authorId: '', length: '', category: ''};
+  let course = { id: '', watchHref: '', title: '', authorId: '', length: '', category: '' };
 
   if (courseId && state.courses.length > 0) {
     course = getCourseById(state.courses, courseId);
   }
 
   return {
-    course: course,
-    authors: authorsFormattedForDropdown(state.authors)
+    course,
+    authors: authorsFormattedForDropdown(state.authors),
   };
 }
 
 function mapDispatchToProps(dispatch) {
   return {
-    actions: bindActionCreators(courseActions, dispatch)
+    actions: bindActionCreators(courseActions, dispatch),
   };
 }
 
