@@ -1,13 +1,14 @@
-/*eslint-disable import/default */
+/* eslint-disable import/default */
 import 'babel-polyfill';
 import React from 'react';
-import {render} from 'react-dom';
+import { render } from 'react-dom';
+import { Provider } from 'react-redux';
+import { Router, browserHistory } from 'react-router';
 import configureStore from './store/configureStore.dev';
-import {Provider} from 'react-redux';
-import {Router, browserHistory} from 'react-router';
 import routes from './routes';
-import {loadCourses} from './actions/courseActions';
-import {loadAuthors} from './actions/authorActions';
+import { loadCourses } from './actions/courseActions';
+import { loadAuthors } from './actions/authorActions';
+import { loadPosts } from './actions/postActions';
 
 // CSS imports
 import '../../node_modules/normalize.css/normalize.css';
@@ -16,10 +17,11 @@ import '../scss/style.scss';
 const store = configureStore();
 store.dispatch(loadCourses());
 store.dispatch(loadAuthors());
+store.dispatch(loadPosts());
 
 render(
   <Provider store={store}>
     <Router history={browserHistory} routes={routes} />
   </Provider>,
-  document.getElementById('app')
+  document.getElementById('app'),
 );
