@@ -2,7 +2,7 @@ import * as types from '../actions/actionTypes';
 import initialState from './initialState';
 
 export default function authorReducer(state = initialState.authors, action) {
-  switch(action.type) {
+  switch (action.type) {
     case types.LOAD_AUTHORS_SUCCESS:
       return action.authors;
 
