@@ -1,12 +1,16 @@
-import {combineReducers} from 'redux';
+import { combineReducers } from 'redux';
 import courses from './courseReducer';
 import authors from './authorReducer';
+import posts from './postReducer';
+import media from './mediaReducer';
 import ajaxCallsInProgress from './ajaxStatusReducer';
 
 const rootReducer = combineReducers({
   courses,
   authors,
-  ajaxCallsInProgress
+  posts,
+  media,
+  ajaxCallsInProgress,
 });
 
 export default rootReducer;
