@@ -1,8 +1,9 @@
-export function authorsFormattedForDropdown (authors) {
-  return authors.map(author=> {
-    return {
-      value: author.id,
-      text: author.firstName + ' ' + author.lastName
-    };
-  });
-}
+const authorsFormattedForDropdown = authors => (
+  authors.map(author => ({
+    value: author.id,
+    text: `${author.firstName} ${author.lastName}`,
+  }),
+  )
+);
+
+export default authorsFormattedForDropdown;
