@@ -1,17 +1,17 @@
 import expect from 'expect';
-import {authorsFormattedForDropdown} from './selectors';
+import authorsFormattedForDropdown from './selectors';
 
 describe('Author Selectors', () => {
   describe('authorFormattedForDropdown', () => {
     it('should return author data formatted for use in a dropdown', () => {
       const authors = [
-        {id: 'cory-house', firstName: 'Cory', lastName: 'House'},
-        {id: 'scott-allen', firstName: 'Scott', lastName: 'Allen'}
+        { id: 'cory-house', firstName: 'Cory', lastName: 'House' },
+        { id: 'scott-allen', firstName: 'Scott', lastName: 'Allen' },
       ];
 
       const expected = [
-        {value: 'cory-house', text: 'Cory House'},
-        {value: 'scott-allen', text: 'Scott Allen'}
+        { value: 'cory-house', text: 'Cory House' },
+        { value: 'scott-allen', text: 'Scott Allen' },
       ];
 
       expect(authorsFormattedForDropdown(authors)).toEqual(expected);
