@@ -1,12 +1,12 @@
-import {createStore, applyMiddleware} from 'redux';
-import rootReducer from '../reducers';
+import { createStore, applyMiddleware } from 'redux';
 import reduxImutableStateInvariant from 'redux-immutable-state-invariant';
 import thunk from 'redux-thunk';
+import rootReducer from '../reducers';
 
 export default function configureStore(initialState) {
   return createStore(
     rootReducer,
     initialState,
-    applyMiddleware(thunk, reduxImutableStateInvariant())
+    applyMiddleware(thunk, reduxImutableStateInvariant()),
   );
 }
