@@ -1,4 +1,5 @@
-if(process.env.NODE_ENV === 'production') {
+/* eslint-disable global-require */
+if (process.env.NODE_ENV === 'production') {
   module.exports = require('./configureStore.prod');
 } else {
   module.exports = require('./configureStore.dev');
