@@ -1,11 +1,11 @@
-import {createStore, applyMiddleware} from 'redux';
-import rootReducer from '../reducers';
+import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
+import rootReducer from '../reducers';
 
 export default function configureStore(initialState) {
   return createStore(
     rootReducer,
     initialState,
-    applyMiddleware(thunk)
+    applyMiddleware(thunk),
   );
 }
