@@ -3,7 +3,7 @@ import path from 'path';
 import open from 'open';
 import compression from 'compression';
 
-/*eslint-disable no-console */
+/* eslint-disable no-console */
 
 const port = 3000;
 const app = express();
@@ -11,11 +11,11 @@ const app = express();
 app.use(compression());
 app.use(express.static('dist'));
 
-app.get('*', function(req, res) {
+app.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, '../dist/index.html'));
 });
 
-app.listen(port, function(err) {
+app.listen(port, (err) => {
   if (err) {
     console.log(err);
   } else {
